Parser collects the created blocks
The parser_cycle and all parser functions returns a list
of tuple each representing a parsed Block:
('block_id', Block_object).

The caller of the parser will manage to add the blocks to the parent_block.
The parend_block is passed as parameter because it could contains
useful informations for the parsing.

Signed-off-by: Davide Valsecchi <valsecchi.davide94@gmail.com>