Added page level collapsing
Create a new way to work on the page tree avoiding the recursion.

-A lot of new method in Page.py to add, remove subpages:
The removal of pages is recursive in the subtree.
The level of added subpages is refreshed along the subtree.

-Added collapsing for page level over content collapsing:
Pages with level higher than a threshold are moved up the tree till the
desired level --> This changes the index not the content.