Plugins can require an instance of TreeExplorer from Renderer
A plugin can request to the Renderer the instance of the TreeExplorer.
The plugin MUST declare two variables:

needs_tree_explorer = True
tree_explorer = None

After the start_rendering method of Renderer the tree_explorer variable
of the plugins is filled.
If a plugin doesn't declare needs_tree_explorer is equal to False.